[Week 02][Day 6] - Pointer
Created a function to calculate and update the circle's area and circumference using pointers (Task 1). Developed an "introduce" function that formats a sentence with all given parameters, utilizing pointers (Task 2). Added a function to insert data into a "buah" slice using pointers and display the contents in a numbered list (Task 3). Lastly, implemented the "tambahDataFilm" function to add map data to the "dataFilm" slice and display it in the specified format (Task 4).